Back End - User Entity Logic => finished

+++ server/controllers/userController.js
@@ -1,4 +1,5 @@
-const { register, login, getAllUsers, toggleConnectionById, deleteProfile, getUserById } = require('../services/userService');
+const { register, login, getAllUsers, toggleConnectionById, deleteProfile, getUserById, editUser } = require('../services/userService');
+const { hasUser } = require('../middlewares/guards');
 const { errorParser } = require('../utils/errorParser');
 
 const userController = require('express').Router();
@@ -24,19 +25,43 @@ userController.get('/details/:id', async (req, res) => {
     }
 });
 
-userController.delete('/delete/:id', async (req, res) => {
-    try {
-        const response = await deleteProfile(req.params.id);
-        res.status(200).json(response);
-    } catch (err) {
-        const errors = errorParser(err);
-        res.status(400).json({ errors });
+userController.put('/edit/:id', hasUser, async (req, res) => {
+    if(req.userData._id == req.params.id) {
+        try {
+            editProfileDto(req)
+            const response = await editUser(req.params.id, {
+                email: req.body.email,
+                firstName: req.body.firstName,
+                lastName: req.body.lastName,
+                profilePic: req.body.profilePic
+            });
+            res.status(200).json(response);
+        } catch (err) {
+            const errors = errorParser(err);
+            res.status(400).json({ errors });
+        }   
+    } else {
+        res.status(400).json({ errors: ['You can edit only your own profile!'] });
     }
 });
 
-userController.post('/connect/:id', async (req, res) => {
+userController.delete('/delete/:id', hasUser, async (req, res) => {
+    if(req.userData._id == req.params.id) {
+        try {
+            const response = await deleteProfile(req.params.id);
+            res.status(200).json(response);
+        } catch (err) {
+            const errors = errorParser(err);
+            res.status(400).json({ errors });
+        }
+    } else {
+        res.status(400).json({ errors: ['You can delete only your own profile!'] });
+    }
+});
+
+userController.post('/connect/:id', hasUser, async (req, res) => {
     try {
-        const response = await toggleConnectionById(req.params.id);
+        const response = await toggleConnectionById(req.params.id, req.userData._id);
         res.status(200).json(response);
     } catch (err) {
         const errors = errorParser(err);
@@ -76,11 +101,15 @@ function registerDto(req) {
         || req.body.lastName == '') {
         throw new Error('all fields are required');
     }
-    if(req.body.email.length < 10) {
-        throw new Error('The email should be at least 10 characters long');
+    const pattern = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    if(!pattern.test(req.body.email)){
+        throw new Error('The email should be in valid format');
     }
-    if(req.body.password.length < 4) {
-        throw new Error('The password should be at least 4 characters long');
+    if(req.body.email.length < 10 || req.body.email.length > 30) {
+        throw new Error('The email should be between 10 and 30 characters long');
+    }
+    if(req.body.password.length < 4 || req.body.password.length > 30) {
+        throw new Error('The password should be be between 4 and 30 characters long');
     }
     if(req.body.firstName.length < 2 || req.body.firstName.length > 20) {
         throw new Error('The firstName should be between 2 and 20 characters long');
@@ -104,5 +133,32 @@ function loginDto(req) {
     }
     if(req.body.password.length < 4) {
         throw new Error('The password should be at least 4 characters long');
+    }
+}
+function editProfileDto(req) {
+    if(req.body.email){
+        const pattern = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+        if(!pattern.test(req.body.email)){
+            throw new Error('The email should be in valid format');
+        }
+        if(req.body.email.length < 10 || req.body.email.length > 30) {
+            throw new Error('The email should be between 10 and 30 characters long');
+        }
+    }
+    if(req.body.firstName){
+        if(req.body.firstName.length < 2 || req.body.firstName.length > 20) {
+            throw new Error('The firstName should be between 2 and 20 characters long');
+        }
+    }
+    if(req.body.lastName){
+        if(req.body.lastName.length < 2 || req.body.lastName.length > 20) {
+            throw new Error('The lastName should be between 2 and 20 characters long');
+        }
+    }
+    if(req.body.profilePic){
+        const pattern = /^https:\/\/res\.cloudinary\.com\//;
+        if(!pattern.test(req.body.profilePic)){
+            throw new Error('Invalid image');
+        }
     }
 }
